requirements.txt
asgiref==3.4.1
boto3==1.20.35
botocore==1.23.35
certifi==2021.10.8
cloudinary==1.28.1
dj-database-url==0.5.0
Django==3.2
django-environ==0.8.1
django-filter==21.1
django-storages==1.12.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
gunicorn==20.1.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
